Queue Using Two Stacks
Oh yeah got the queue using two stack just by readning the strategy and not looking at any code. It wasn't perfect but improving. Also HakerRank does't passes some tests but it might be due to the loops but we can improve later.

We got a working solution, which is what matters when trying to find a solutions and using brute force