Getting Ice Cream Parlor to work efficiently
Oh boy this was a hard one. The Java example were of some help but I needed to make sure I adjusted to C#. My code was throwing an exception because a key already existed in the dictionar, and in C# the dictionary element doesn't update like in other languages.